refactor: move version filter function to utils
